Create video_utils module
Video utils now functional for deployment. Some updates to config_utils included.
